Fix script on MacOS
It would be nice to add --pure to the nix-shell invocation as well, and
pull in git and nix as the other shell's dependencies. This will pull in
a nix without flake features though which the script requires. Choosing
not to spend the time right now diving deeper into this rabbit hole.